Fracking: Reduce size of the model for speed
Model is 1/4 of size and loads and runs faster.

Note: There is an implicit relationship between the size of
the model and many of the properties (speed etc.). Some of
these have been reduced (depth of layers is now a %), but
others still remain, which means they will have to be changed if
model size is changed.

Since model size is a part of the model, we should consider making
model size part of the model definition, not just specified
in the html page.

[#54298936]