Climate: Move controls to common script; add jquery sliders
Controls have some minimal branching based on the model type
passed in from the html page, but for the most part it
simply depends on jquery functions not being called when
the selector doesn't exist.

Added in jquery sliders because input-range elements are not
well supported, especially in firefox.

Added some minimal styling for jquery sliders, making them look
a little more lab-like