Climate: Create a speed slider for var 1; start model slower
Note that the speed slider only works when async is false. This
attribute can only be set either at the start or after a pause
and restart. So if we want to add this to models that need to be
fast (i.e. with async true), the slider will need to do a quick
pause and restart of the model in order to set it.