Land Mgmt: keep grass roots from prematurely dying
[#70095564]

Problem was that the formula that was intended to randomly vary the growth period lengths was generating NaNs about half the time (whenever Uniform[-v, v] < 0, it tried to add positive and negative infinities). Spaghetti logic elsewhere in the code checks for a period length equal to Infinity in order to determine whether a plant is supposed to live indefinitely or not.