air pollution: rm awkward baseHeading property
[#67429718]

We want to keep track of the slightly drifting long-term heading of the pollution particles, but also allow the particles to drift upward because of temperature, or to the left or right due to the wind.

The previous solution was to use AgentScript’s forward() method for every step of the particles’ motion, but the forward() method requires the direction to be kept in the “heading” property of the agent. Somewhat awkwardly, the code kept a separate property, “baseHeading” to store the long-term heading of the particle, while using “heading” as basically an argument to forward().

As a simplification I use “heading” to store the long-term heading, and use setXY() instead of forward() for motions which don’t require knowing the long-term direction of the particle’s motion.